CI: Add step to deploy docs from
This step will be run each time on master and uses a new branch
`gh-pages` which contains a subtree containing just the built docs.

`git` is configured with some placeholder details, and the commit for
each deploy just gives the SHA on master the docs were built from.

The docs should then be available at [1]

[1] https://matthewhughes934.github.io/pystatsd/