CI: Avoid copying extra virtualenvs around
I want to copy at least one virtual env so that I can build once and:

* test
* lint
* build docs

which requires a virtualenv for a single Python job. However, the
current setup builds and stores a virtualenv for each Python version
(these can be rather large, ~45mb) which makes downstream jobs slower.
Instead, split up jobs so that there's one 'build and test' job for all
Python versions except one. For this final job, separate the 'build'
step so we can store the virtualenv and pass it around as before.